elabs/app/core/requirements.txt

pymssql==2.1.5   # for mac
fire
requests
python-dateutil

